+++ CHANGES.rst
@@ -10,6 +10,7 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Add thresholds to charts.
 
+- Speed up angular: set debugInfoEnabled of angular compiler to false.
 
 Release 3.0.13 (2016-4-29)
 ---------------------